Reorganize high-level functions.  Use these in the legacy wrapper as well.
